test: added test for Calculate function
