+++ cmd/root_test.go
@@ -1,20 +1,8 @@
 package cmd
 
 import (
-	"log"
 	"testing"
 )
-
-func TestCalculateCost(t *testing.T) {
-	expected := 0.000021333376
-	actual := calculateCost(1280.0, 1024)
-
-	log.Println(actual)
-
-	if actual != expected {
-		t.Errorf("expected %f, got %f", expected, actual)
-	}
-}
 
 func TestMemorySortedList(t *testing.T) {
 	expected := []int{128, 256, 384, 512}
